+++ osu.Game.Rulesets.Catch/Objects/Drawables/IHasFruitState.cs
@@ -1,0 +1,15 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Bindables;
+
+namespace osu.Game.Rulesets.Catch.Objects.Drawables
+{
+    /// <summary>
+    /// Provides a visual state of a <see cref="Fruit"/>.
+    /// </summary>
+    public interface IHasFruitState : IHasCatchObjectState
+    {
+        Bindable<FruitVisualRepresentation> VisualRepresentation { get; }
+    }
+}
